Added regex process and implemented on identifier

+++ src/star/Debug.cc
@@ -2,13 +2,13 @@
 #include <iostream>
 #include "ColorToken.hh"
 
-void star::Debug::Report(uint32_t line, const std::string& where, const std::string& message)
+void star::Debug::Report(size_t line, const std::string& where, const std::string& message)
 {
     s_HadError = true;
     std::cerr << ColorToken::s_Red << "Error at line: " << line << ", on: " << where << "\n" << message;
 }
 
-void star::Debug::Error(uint32_t line, const std::string& message)
+void star::Debug::Error(size_t line, const std::string& message)
 {
     Report(line, "", message);
 }
