+++ src/star/Debug.hh
@@ -1,7 +1,6 @@
 #pragma once
 
 #include "StarMacro.hh"
-#include <cstdint>
 #include <string>
 
 namespace star
@@ -9,12 +8,12 @@ namespace star
     class STAR_API Debug
     {
     private:
-        static void Report(uint32_t line, const std::string& where, const std::string& message);
+        static void Report(size_t line, const std::string& where, const std::string& message);
 
         inline static bool s_HadError = false;
 
     public:
-        static void Error(uint32_t line, const std::string&);
+        static void Error(size_t line, const std::string&);
         static bool HadError();
     };
 }
